data_format_translation_server_homework/python-server/pyproject.toml
[tool.poetry]
name = "python-server"
version = "0.1.0"
description = ""
authors = ["aleksanderminchev <83026219+aleksanderminchev@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
bs4 = "^0.0.1"
PyYAML = "^6.0"
lxml = "^4.9.1"
xmltodict = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
